move qr code into its own widget
